Team08_DontCare/Final Round Submission/requirements.txt
aiohttp
aiosignal
altair
annotated-types
attrs
azure-cognitiveservices-speech
blinker
cachetools
certifi
charset-normalizer
click
chromadb
colorama
dataclasses-json
frozenlist
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
httplib2
idna
jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
markdown-it-py
markupsafe
marshmallow
mdurl
multidict
mypy-extensions
mysql-connector-python
numpy
orjson
packaging
pandas
pillow
proto-plus
protobuf
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pygments
pyparsing
python-dateutil
python-dotenv
pymysql
pytz
pyyaml
referencing
requests
rich
rpds-py
rsa
six
smmap
sqlalchemy
streamlit
tenacity
toml
toolz
tornado
tqdm
typing-extensions
typing-inspect
tzdata
uritemplate
urllib3
watchdog
yarl
langchain_google_genai
